make the test more comprehensive
