--- dev/integration_tests/android_engine_test/lib/hcpp/platform_view_transform_rotation_main.dart
@@ -1,111 +1,0 @@
-// Copyright 2014 The Flutter Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import 'dart:convert';
-import 'dart:math' as math;
-
-import 'package:android_driver_extensions/extension.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter/gestures.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_driver/driver_extension.dart';
-
-import '../src/allow_list_devices.dart';
-
-void main() async {
-  ensureAndroidDevice();
-  enableFlutterDriverExtension(
-    handler: (String? command) async {
-      return json.encode(<String, Object?>{
-        'supported': await HybridAndroidViewController.checkIfSupported(),
-      });
-    },
-    commands: <CommandExtension>[nativeDriverCommands],
-  );
-
-  // Run on full screen.
-  await SystemChrome.setEnabledSystemUIMode(SystemUiMode.immersive);
-  runApp(const MainApp());
-}
-
-final class MainApp extends StatefulWidget {
-  const MainApp({super.key});
-
-  @override
-  State<MainApp> createState() => _MainAppState();
-}
-
-class _MainAppState extends State<MainApp> with SingleTickerProviderStateMixin {
-  late double angle;
-
-  void _incrementAngle() {
-    setState(() {
-      angle = angle + 0.5;
-    });
-  }
-
-  @override
-  void initState() {
-    angle = 0;
-    super.initState();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final Widget widget = Transform.rotate(
-      angle: angle * math.pi,
-      child: Stack(
-        alignment: Alignment.center,
-        children: <Widget>[
-          TextButton(
-            key: const ValueKey<String>('IncrementAngle'),
-            onPressed: _incrementAngle,
-            child: const SizedBox(width: 300, height: 500, child: ColoredBox(color: Colors.green)),
-          ),
-          const SizedBox(
-            width: 200,
-            height: 400,
-            child: _HybridCompositionAndroidPlatformView(
-              viewType: 'blue_orange_gradient_platform_view',
-            ),
-          ),
-        ],
-      ),
-    );
-
-    return MaterialApp(debugShowCheckedModeBanner: false, home: Center(child: widget));
-  }
-}
-
-final class _HybridCompositionAndroidPlatformView extends StatelessWidget {
-  const _HybridCompositionAndroidPlatformView({required this.viewType});
-
-  final String viewType;
-
-  @override
-  Widget build(BuildContext context) {
-    return PlatformViewLink(
-      viewType: viewType,
-      surfaceFactory: (BuildContext context, PlatformViewController controller) {
-        return AndroidViewSurface(
-          controller: controller as AndroidViewController,
-          gestureRecognizers: const <Factory<OneSequenceGestureRecognizer>>{},
-          hitTestBehavior: PlatformViewHitTestBehavior.transparent,
-        );
-      },
-      onCreatePlatformView: (PlatformViewCreationParams params) {
-        return PlatformViewsService.initHybridAndroidView(
-            id: params.id,
-            viewType: viewType,
-            layoutDirection: TextDirection.ltr,
-            creationParamsCodec: const StandardMessageCodec(),
-          )
-          ..addOnPlatformViewCreatedListener(params.onPlatformViewCreated)
-          ..create();
-      },
-    );
-  }
-}
